refactor: Firebase設定ファイルをconfigディレクトリへ配置
