+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:roadmap/firebase_options.dart';
+import 'package:roadmap/core/config/firebase_options.dart';
 import 'package:roadmap/presentation/routes/app.dart';
 
 Future<void> main() async {
